Fix up namespaces and move tool window
