+++ src/Education.UnitTests/Commands/RuleHelpWindowCommandTests.cs
@@ -29,6 +29,7 @@ using SonarLint.VisualStudio.TestInfrastructure;
 using SonarLint.VisualStudio.Education.Commands;
 
 using ThreadHelper = SonarLint.VisualStudio.TestInfrastructure.ThreadHelper;
+using SonarLint.VisualStudio.Education;
 
 namespace Education.UnitTests.Commands
 {
